pkgm outdated
Really we need to record what constraint the user requested when installing as well as what packages were explicitly installed in order to do a better job here.

Currently we assume any ^x that is newer counts as outdated provided it is not a package that is a dependency with a tighter constraint range.

This also ties into how we handle multiple installs, which currently we do, eg. you can have openssl^1 and openssl^3 both installed though this works by luck rather than by design. If you have both then we need to state that both ^1 and ^3 are outdated, this naive implementation will actually error currently for this case.